Fix missing Wireserver IP configuration in Wireserver Client (#2271)
* Add coverage around GetInterfaces in WS client

There's a bug with the wireserver client, in that the Wireserver IP from
the CNS config isn't respected. This adds coverage so that it becomes
safer to make changes here.

* Allow Wireserver client to use a configurable Host

This allows the Wireserver client to use a configurable host so that
this could, theoretically be mocked. As it currently stands, it's an
impediment to running CNS locally, since it uses the real Wireserver IP.

* Add WireserverIP config into Wireserver Client

The Wireserver client now accepts a hostport configuration option so
that its requests can be redirected. This plumbs the config for the
WireserverIP to it so that it can be redirected during test.

* fix: replace nil with http.NoBody

Signed-off-by: Evan Baker <rbtr@users.noreply.github.com>

---------

Signed-off-by: Evan Baker <rbtr@users.noreply.github.com>
Co-authored-by: Tim Raymond <traymond@microsoft.com>